requirements.txt
streamlit
google-generativeai
python-dotenv
ipython
Pillow